Added example item
Added an example item in the tomorrow field. Eventually, I hope to be
able to save these and automatically add a new one every time the above
one is filled. Either that, or we would need to build functions for
adding and removing them. Actually, we probably need to build a remove
function regardless of an add.